pyproject.toml
[tool.poetry]
name = "hello"
version = "0.1.0"
description = ""
authors = ["David2261 <bulatnasirov2003@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
sentry-sdk = "^1.9.10"
djangorestframework = "^3.14.0"
django-grappelli = "^3.0.3"
django-ckeditor = "^6.5.1"
django-tinymce = "^3.5.0"
whitenoise = "^6.2.0"
Pillow = "^9.2.0"
pytest-django = "^4.5.2"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
flake8-django = "^1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

